feat(openapi): support Path Item references
